+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * deprecated `IteratorListAdapter`
 * fixed a bug that disabled json if ujson was not installed
+* write_to_file_json has kwargs